move Twitter fns to libs/model
