+++ libs/model/src/services/index.ts
@@ -6,3 +6,4 @@ export * from './snapshot';
 export * as stakeHelper from './stakeHelper';
 export * as tokenBalanceCache from './tokenBalanceCache';
 export * from './tokenBalanceCache/types';
+export * from './twitter';
