+++ libs/model/src/services/twitter/index.ts
@@ -1,0 +1,5 @@
+export * from './api/getMentions';
+export * from './api/getTweets';
+export * from './twitter.config';
+export * from './types';
+export * from './utils';
