+++ packages/commonwealth/server/config.ts
@@ -1,7 +1,7 @@
 import { config as adapters_config } from '@hicommonwealth/adapters';
 import { configure } from '@hicommonwealth/core';
 import { config as model_config } from '@hicommonwealth/model';
-import { ChainBase, TwitterBotName } from '@hicommonwealth/shared';
+import { ChainBase } from '@hicommonwealth/shared';
 import { z } from 'zod';
 
 const {
@@ -27,12 +27,9 @@ const {
   DEV_MODULITH,
   ENABLE_CLIENT_PUBLISHING,
   EVM_CE_LOG_TRACE,
-  TWITTER_WORKER_POLL_INTERVAL,
   CACHE_GET_COMMUNITIES_TRENDING_SIGNED_IN,
   CACHE_GET_COMMUNITIES_TRENDING_SIGNED_OUT,
   CACHE_GET_COMMUNITIES_JOIN_COMMUNITY,
-  TWITTER_ENABLED_BOTS,
-  TWITTER_APP_BEARER_TOKEN,
 } = process.env;
 
 const DEFAULTS = {
@@ -118,18 +115,6 @@ export const config = configure(
     },
     DEV_MODULITH: DEV_MODULITH === 'true',
     ENABLE_CLIENT_PUBLISHING: ENABLE_CLIENT_PUBLISHING === 'true',
-    TWITTER: {
-      WORKER_POLL_INTERVAL: (() => {
-        if (TWITTER_WORKER_POLL_INTERVAL)
-          return parseInt(TWITTER_WORKER_POLL_INTERVAL, 10);
-        else if (model_config.APP_ENV === 'local')
-          return DEFAULTS.TWITTER_WORKER_POLL_INTERVAL;
-        else return 0;
-      })(),
-      ENABLED_BOTS:
-        (TWITTER_ENABLED_BOTS?.split(',') as TwitterBotName[]) || [],
-      APP_BEARER_TOKEN: TWITTER_APP_BEARER_TOKEN,
-    },
     CACHE_TTL: {
       GET_COMMUNITIES_TRENDING_SIGNED_IN:
         CACHE_GET_COMMUNITIES_TRENDING_SIGNED_IN
@@ -227,15 +212,6 @@ export const config = configure(
     }),
     DEV_MODULITH: z.boolean(),
     ENABLE_CLIENT_PUBLISHING: z.boolean(),
-    TWITTER: z
-      .object({
-        APP_BEARER_TOKEN: z.string().optional(),
-        WORKER_POLL_INTERVAL: z.number().int().gte(0),
-        ENABLED_BOTS: z.array(z.nativeEnum(TwitterBotName)),
-      })
-      .refine(
-        (data) => !(data.ENABLED_BOTS.length > 0 && !data.APP_BEARER_TOKEN),
-      ),
     CACHE_TTL: z.object({
       GET_COMMUNITIES_TRENDING_SIGNED_IN: z.number(),
       GET_COMMUNITIES_TRENDING_SIGNED_OUT: z.number(),
