+++ packages/commonwealth/server/workers/twitterWorker/utils.ts
@@ -1,49 +1,12 @@
-import { models } from '@hicommonwealth/model';
-import { events, Tweet } from '@hicommonwealth/schemas';
-import { TwitterBotName } from '@hicommonwealth/shared';
+import {
+  models,
+  TwitterBotConfig,
+  TwitterBotConfigs,
+  twitterMentions,
+} from '@hicommonwealth/model';
+import { Tweet } from '@hicommonwealth/schemas';
 import { QueryTypes, Transaction } from 'sequelize';
 import { z } from 'zod';
-import { config } from '../../config';
-
-export const TwitterBotConfigs = {
-  [TwitterBotName.MomBot]: {
-    name: TwitterBotName.MomBot,
-    // TODO: update
-    username: 'mombot',
-    twitterUserId: '1337',
-    bearerToken: config.TWITTER.APP_BEARER_TOKEN,
-  },
-  [TwitterBotName.ContestBot]: {
-    name: TwitterBotName.ContestBot,
-    // TODO: update
-    username: 'contestbot',
-    twitterUserId: '1338',
-    bearerToken: config.TWITTER.APP_BEARER_TOKEN,
-  },
-  [TwitterBotName.Common]: {
-    name: TwitterBotName.Common,
-    username: 'commondotxyz',
-    twitterUserId: '1005075721553932288',
-    bearerToken: config.TWITTER.APP_BEARER_TOKEN,
-  },
-} as const;
-
-export type TwitterBotConfig =
-  (typeof TwitterBotConfigs)[keyof typeof TwitterBotConfigs];
-
-export type twitterMentions =
-  | Array<{
-      event_name: 'TwitterMomBotMentioned';
-      event_payload: z.infer<typeof events.TwitterMomBotMentioned>;
-    }>
-  | Array<{
-      event_name: 'TwitterContestBotMentioned';
-      event_payload: z.infer<typeof events.TwitterContestBotMentioned>;
-    }>
-  | Array<{
-      event_name: 'TwitterCommonMentioned';
-      event_payload: z.infer<typeof events.TwitterCommonMentioned>;
-    }>;
 
 export function createMentionEvents(
   twitterBotConfig: TwitterBotConfig,
